Add poolMaxIdleTime configuration option to TCP appenders
When connections remain open without being used, they may be
considered stale by the server and closed. This can lead to log messages
failing with

java.lang.SocketException: broken pipe

This commit introduces a maximum 'idle time' for pooled connections. If
a connection has not been used within that time, it will be considered
stale and not reused. By default, this new feature is disabled.